Create an initial backup in a backup loop.
Previously, the backup was created only when the time since the
last backup exceeded the backup interval, or on a specific hour.
With this change, the backup is also created when the cluster is
started initially, but not during the failover. The new code also
minimizes the number of times it has to query S3.

In addition, more sane parameter values for the backup hour/interval
are added. Those parameters are now part of the Dockerfile, so that
one can modify them when the cloudformation template is created with
senza.