Command-line client for Spilo
This command line client is to aid in connecting to spilo clusters running inside a vpc.
It allows to create a tunnel and connect using a single command.

The healthcheck will only connect to the governor http port of the specified host and
return its output.

The tunnel will create a tunnel for you and return a connect string.