Spilo command line interpreter: Harden some options
If a service is specified, it will now be appended to psql to be executed. This removes the need
to pass on individual options on the command line, the most obvious improvement is that passwords can now
be specified in a pg_service file, and they will not be visible on the command line. (This still requires the
end user to ensure his or her pg_service file has strict permissions).

The tunnel is now created and instead of waiting an arbitrary amount, we wait until we have established connectivity or
we timeout.