Explicitly set keepalives.
To ensure we keep the connection alive to the ELB, we send a keepalive after 15 minutes.
The default of the OS is normally 2 hours, the maximum for the ELB is 3600 seconds.

This addresses issue #25 partly, by ensuring the connection between the ELB and PostgreSQL is
kept alive.

This also requires the ELB connection timeout to increase.