Rename Spilo members to match the AWS instance id.
For easier cross-referencing between logfiles, AWS consoles and Spilo, it is
usefull to be able to see which EC2 instance id belongs to Spilo.
The previous identifier used the Docker hostname, which does not convey
a lot of information in debugging.

To ensure replication slots can still be created, the dash needs to be transformed
into an underscore.