Allow patroni_wait to execute a specified command
For supervisor jobs it is very useful to be able to
specify 1 command to start a job.

With Patroni/Spilo it is sometimes necessary to wait for a process to
be available before starting the job.

These changes allow you to prepend your command with a wait on Patroni
with an optional timeout.