Reduce the pool size for backup
The default pool size for WAL-E is 4, we've seen some Spilo appliances
failover during the backup. During the backup, the CPU utilization was very
high and the lock in etcd was not updated on time.

To reduce stress on systems with less cores, we will reduce the pool size
if the number of cores is small