Introduce an stunnel to connect to an ldaps
This feature allows us to connect to an ldap server that only
allows a direct SSL connection. This however is not supported for
PostgreSQL, see [1]. By adding an stunnel process if and only
if an LDAP_URL is specified, we ensure only those Spilo's that
require this stunnel will have it configured.

A dynamic configuration would be nice, currently it is static due
to the Certificate Chain having only specific certificates.
Any future dynamic configuration needs to setup a secure way
of setting these certificates.

[1] https://www.postgresql.org/docs/current/static/auth-methods.html#AUTH-LDAP