Fix setting permissions on arbitrary mounts.
Normally, when PGDATA is located on a mounted volume,
we need to make sure that the volume root is accessible
to the postgres user. Since we cannot guarantee that, we
create another directory, called pgroot, and reassign
permissins for it. The actual PGDATA is then created inside
the PGROOT.

Also pass GOOGLE_APPLICATION_CREDENTIALS to WAL-E.