Configuration can be set with $SPILO_CONFIGURATION.
Previously, Spilo used PATRONI_CONFIGURATION in order to read the
non-default configuration for Patroni (the default one is hard-coded
in the configure_spilo.py. Because the envionment variable with the
same name is also used by Patroni for a completely different purpose,
it lead to a conflict that was hard to debug. Therefore, we need to
rename it, but in order to keep the compatibility with the already
deployed clusters, both the old and the new (SPILO_CONFIGURATION)
names will be accepted by the Docker image with the preference for
the new one if both are set. That will allow us to change the
deployment configuration to use the new name without breaking already
deployed clusters in the case we will need to update their configuration
on the fly.