Some optimisations around restore_command = wal-e
1) wal-e wal-fetch is incredibly slow
2) When we start postgres it will call the restore_command for every
xlog segment needed to reach consistent state even if there is a valid
file in the pg_xlog dicrectory.
3) Even if the wal-e is prefetching wal segments it it is rather
expencive to move them into pg_xlog by wal-e.

All these facts are making useless streaming of xlogs with pg_basebackup,
because afterwards it will call restore_command anyway :(

In order to optimize this process we will stream xlogs not into default
pg_xlog but into separate directory and create a wrapper restore_command
script, which will:
Check if there is xlog already available on disk either in directory
created by pg_basebackup or in the pg_xlog/.wal-e/prefetch directory.
If the file is there - just move it. If it is not - call wal-e with
usual options.