Install pg_repack
In order to be able to install it for 9.3 and 9.4 I had to change the
way how we install postgres packages and 3rd party extensions.
Now they are install from the same layer, starting from postgres 9.3 and
up to 9.6. It's necessary, because otherwise it was not possible to
build pg_repack for 9.3 and 9.4. pg_repack require matching version of
libpq-dev but by default ubuntu is installing only the latest one.