Don't override Name tags of EBS volumes
Sometime we rely on Name tag of the data EBS volume, which is different from
the usual 'spilo_{cluster_name}' format, for example, when attaching multiple
volumes and making a RAID on top of them.

This will ensure that any pre-defined Name tags are kept intact.  The instance
and role tags are updated in any case.