Make post_init.sh script idempotent (#236)
The second run will not produce errors but rather will fix some discrepancies between current and desired state.
In addition to that it makes zmon_utils compatible with postgres >= 10.

On the next step we can call this script after every "promote" event. It will eliminate the need to manually deploy future changes.