Execute pg_upgrade after custom bootstrap if major version doesn't match
Before the actuall call of pg_upgrade the maybe_pg_upgrade script starts
the old postgres binary, waits until pg_is_in_recovery() starts
returning `False` and does a clean shutdown.

In addition to that, the post_init.sh script calls vacuumdb to analyze
all databases in the cluster. In case of bootstrapping the new cluster
(with or without pg_upgrade) it will do `--analyze-in-stages` and after
the failover it will do `--analyze-only` in background.