Define RESTAPI_CONNECT_ADDRESS environment variable (#640)
Spilo by default uses IP for the property ```restapi.connect_address```. This is a problem when Patroni REST API is configured in SSL with client authentication because hostname validation is performed. Due to the dynamic nature of Pod, it's impossible to create SSL certificates with the right IP into SAN.
This PR solves the problem with the introduction of the property: REST_API_CONNECT_ADDRESS.
You can configure it, for example, like this:
```
        - name: RESTAPI_CONNECT_ADDRESS
          value: "$(POD_NAME).<service name>"
``` 

so that it uses the <pod name>.<service name> stored in CoreDNS.