Add retry to launch.sh for configure_spilo.py
Add a small delay and single retry for /scripts/configure_spilo.py from /launch.sh.

It is possible that a large cluster with iptables-enforced network policies could have a ever so brief delay in enforcing network policies.  In my testing it was possible to trigger in a ~150 node cluster with sufficient regularity to cause additional admin work.  Having a single retry after a few seconds alleviates the issue so we don't have to destroy/recreate the pod (and if it still fails after the retry, we'll destroy/recreate the pod anyways, so the only cost is a brief delay upon initial failure).