Fall back in case wal-g restore_command hangs
The `restore_command` might not terminate (or at least run for a
very long time). This is the case, for instance, if wal-g is
enabled and it cannot reach the archive location.
See: https://github.com/wal-g/wal-g/issues/1198

This commit adds a configurable timeout for it, disabled by
default. If the timeout triggers, then the `restore_command`
terminates with exit code 124. PostgreSQL will next try to restore
WAL from the `pg_wal` directory, then from the primary server (if
configured) and, if that fails, then again from the archive
location.
See: https://www.postgresql.org/docs/current/warm-standby.html#STANDBY-SERVER-OPERATION

This way, a standby server can, in many (but not all) cases, start
even if the archive location is down. In order to enable this
behavior, the user can set the `WAL_RESTORE_TIMEOUT` environment
variable to a nonzero integer.