Add new environment variable PGCONNECT_ADDRESS to set the postgresql.connect_address
Spilo by default uses IP for the property ```postgresql.connect_address```. This is a problem when Patroni REST API is configured in SSL with client authentication because hostname validation is performed or behind NAT. This PR solves the problem with the introduction of environment variable: PGCONNECT_ADDRESS. You can configure it with the external IP of the spilo or dns name of spilo and the replicas will use the address PGCONNECT_ADDRESS:PGPORT when connecting the primary.